requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
PyMuPDF>=1.25.0
python-pptx>=0.6.21
deep-translator>=1.11.0
Pillow>=10.0.0
